pyproject.toml
[project]
name = "multimodal-coordination"
version = "0.1.0"
description = "Detecting coordinated inauthentic behaviour via multimodal machine learning and network science"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }

dependencies = [
    "networkx>=3.0",
    "pandas>=1.5",
    "numpy>=1.24",
    "sentence-transformers>=2.2",
    "torch>=2.0",
    "Pillow>=9.0",
    "leidenalg>=0.10",
    "igraph>=0.11",
    "python-igraph>=0.11",
    "tqdm>=4.0",
    "coordination-network-toolkit",
]

[project.scripts]
multimodal-coordination = "multimodal_coordination.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["multimodal_coordination"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ipykernel",
]
